requirements.txt
asgiref==3.7.2
bleach==6.1.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-admin-geomap==1.6
django-constance==3.1.0
django-import-export==3.3.3
django-picklefield==3.1
django-summernote==0.8.20.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
psycopg2==2.9.9
